+++ apps/avatax/src/modules/public-log-drain/public-log-drain.ts
@@ -1,0 +1,11 @@
+import { PublicLog } from "./public-events";
+
+export interface PublicLogDrain {
+  emitLog(log: PublicLog): Promise<void[]>;
+  getTransporters(): LogDrainTransporter[];
+  addTransporter(transporter: LogDrainTransporter): void;
+}
+
+export interface LogDrainTransporter {
+  emit(log: PublicLog): Promise<void>;
+}
